Fix check-vendors intermediate rules
Dune seems to get confused and runs the rules for `upstream_p256_*.h`
when running `dune build`. For some reason dune feels like it needs
those `.h` files. I'm guessing it has no way to tell which `.h` files it
need in advance so I'm not sure we should report this upstream.

In the meantime I renamed the targets to remove the `.h` extension and
it fixes the issue.