Fix Bigarray linking issue by ensuring cstruct < 0.5.0
Cstruct 5.0.0 dropped its dependency over `bigarray` in favor of
`bigarray-compat` transitively breaking `fiat`. This could be fixed
by also using `bigarray-compat` instead of `bigarray`, which would be
the right thing to do, but because `asn1-combinators` also transitively
depends on `bigarray` we'd still end up breaking any binary built using
fiat.

This is a temporary solution so I'll open an issue as a remainder that
we need to upgrade once `asn1-combinators` does.